Pulled out functions and imports
